+++ bugsnag-plugin-android-ndk/src/main/java/com/bugsnag/android/NdkPlugin.kt
@@ -16,6 +16,8 @@ internal class NdkPlugin : Plugin {
     private external fun enableCrashReporting()
     private external fun disableCrashReporting()
 
+    private external fun getBinaryArch(): String
+
     private var nativeBridge: NativeBridge? = null
     private var client: Client? = null
 
@@ -46,6 +48,7 @@ internal class NdkPlugin : Plugin {
             true
         }
         if (libraryLoader.isLoaded) {
+            client.setBinaryArch(getBinaryArch())
             nativeBridge = initNativeBridge(client)
         } else {
             client.logger.e(LOAD_ERR_MSG)